+++ lib/element_params/sub/button/radio_btn_params.dart
@@ -1,0 +1,15 @@
+import 'checkbox_params.dart';
+
+class RadioBtnParamsWrapper {
+  RadioBtnParams p;
+
+  /// * [p] : The parameter set.
+  RadioBtnParamsWrapper(this.p);
+}
+
+class RadioBtnParams {
+  SelectableIconBtnParams? enableParams;
+  SelectableIconBtnParams? disableParams;
+  int? selectedIndex;
+  List<void Function(int? selectedIndex)?>? callbacks;
+}