+++ lib/element_params/sub/structure/table_row_params.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class TableRowParamsWrapper {
+  TableRowParams p;
+
+  /// * [p] : The parameter set.
+  TableRowParamsWrapper(this.p);
+}
+
+class TableRowParams {
+  LocalKey? key;
+  Decoration? decoration;
+  List<Widget>? children;
+}